fixed warnings, fixed drag and drop bug

+++ compiler/compiler.cpp
@@ -310,7 +310,7 @@ namespace Aseba
 					else
 					{
 						int index = ASEBA_EVENT_LOCAL_EVENTS_START - eventId;
-						if (index < targetDescription->localEvents.size())
+						if (index < (int)targetDescription->localEvents.size())
 							dump << "event " << targetDescription->localEvents[index].name << ": ";
 						else
 							dump << "unknown local event " << index << ": ";
